+++ app/controllers/events_controller.rb
@@ -81,7 +81,11 @@ class EventsController < ApplicationController
   def event_params
     params.require(:event).permit(
       :name, :slug, :description, :position, :archived, :only_unblinded,
-      design_hashes: [:design_id, :handoff_enabled]
+      design_hashes: [
+        :design_id, :handoff_enabled, :duplicates, :requirement,
+        :conditional_event_id, :conditional_design_id, :conditional_variable_id,
+        :conditional_operator, :conditional_value
+      ]
     )
   end
 end