+++ db/migrate/20170309210803_add_duplicates_to_event_designs.rb
@@ -1,0 +1,5 @@
+class AddDuplicatesToEventDesigns < ActiveRecord::Migration[5.0]
+  def change
+    add_column :event_designs, :duplicates, :string, null: false, default: 'highlight'
+  end
+end