+++ CHANGELOG.md
@@ -29,6 +29,7 @@
     100% complete as opposed to 0% complete
       - This change helps mark handoff designs complete that only include
         sections with descriptions
+  - Improved the autocomplete for variable values on the sheets index
 - **Site Changes**
   - Default site short name is now properly removed when the default site is
     renamed during the project setup process